+++ src/app/event/event.page.ts
@@ -97,6 +97,7 @@ export class EventPage implements OnInit, OnDestroy {
   mapSubtitle = '';
   campDescription = '';
   locationInfo = '';
+  campImage: string | undefined;
   prevDisabled = false;
   nextDisabled = false;
   routerOutlet: IonRouterOutlet = inject(IonRouterOutlet);
@@ -214,6 +215,7 @@ export class EventPage implements OnInit, OnDestroy {
     const camp = await this.db.findCamp(this.event?.hosted_by_camp!);
     if (camp) {
       this.campDescription = camp.description!;
+      this.campImage = camp.imageUrl;
       this.isOpen = true;
     }
     this._change.detectChanges();